--- Sources/Alembic.h
@@ -1,17 +1,0 @@
-//
-//  Alembic.h
-//  Alembic
-//
-//  Created by Ryo Aoyama on 4/3/16.
-//  Copyright © 2016 Ryo Aoyama. All rights reserved.
-//
-
-@import Foundation;
-
-//! Project version number for Alembic.
-FOUNDATION_EXPORT double AlembicVersionNumber;
-
-//! Project version string for Alembic.
-FOUNDATION_EXPORT const unsigned char AlembicVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Alembic/PublicHeader.h>
